Add app models

+++ APISistemaTickets/Data/Models/Auth/User.cs
@@ -1,10 +1,11 @@
+using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
 namespace APISistemaTickets.Data.Models.Auth;
 
 public class User
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string FirstName { get; set; }
     public string LastName { get; set; }
     public string Username { get; set; }
